Engine no longer uses mesh names.
As the engine is for a voxel engine, and meshes will never be re-used,
the system of using mesh names was nothing more than a way to slow down
the engine due to std::string's overhead. As a result, moving to use a
simple boolean to check if a chunk is set to be rendered or not is more
than enough to check if the engine should display the chunk or not. This
resulted in a massive improvement and shorter lag spikes when moving.

Moved chunk rendering code out of the player movement check. Previously,
chunks that weren't immediately generated would only start to render
after the player moved out of the current chunk. Now all chunks are to
be displayed even if the player remains still.